requirements.txt
Flask
python-docx
docx2pdf
openai
gunicorn
json